requirements.txt
aniso8601==9.0.1
blessed==1.20.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-RESTful==0.3.10
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-editor==1.0.4
python-gitlab==3.15.0
pytz==2023.3
readchar==4.0.5
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
